Fixed like button not working and moved some inline styling
The liked button is fixed by first creating a new prop for the component based on the liked state. Before unmounting, we'll check if the current liked is the same as the original one and if it isn't call the handleLike function. The liked state is only for the display of the heart and to change it using onPress